Refactor: Response, Request, Dto 구분을 위한 vo 패키지 추가
